requirements.txt
# PDF handling
pdfplumber
PyMuPDF  # fitz

# OCR
pytesseract
Pillow

# Text processing
regex
nltk

# Image utilities
opencv-python-headless  # safer for servers/Colab

# For future RAG pipeline
torch
torchvision
transformers
huggingface_hub
sentence-transformers
langchain
chromadb
tqdm
numpy
scikit-learn

# Optional (for Groq/OpenAI-style API calling)
openai
requests

# DuckDuckGo search for fallback
duckduckgo-search

# If working in Colab
ipython
